# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Softcurse LiveScriptor IDE API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: projects
    description: Project management
  - name: files
    description: File system operations
  - name: terminal
    description: Terminal execution
  - name: ai
    description: AI assistant
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all projects
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a new project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectRequest"
      responses:
        "201":
          description: Created project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"

  /projects/{projectId}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get project by ID
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Project details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Project not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /projects/{projectId}/files:
    get:
      operationId: listFiles
      tags: [files]
      summary: List files in a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
        - name: path
          in: query
          schema:
            type: string
            default: "/"
      responses:
        "200":
          description: File tree
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FileNode"
    post:
      operationId: createFile
      tags: [files]
      summary: Create a file or directory
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFileRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileNode"

  /projects/{projectId}/files/content:
    get:
      operationId: getFileContent
      tags: [files]
      summary: Get file content
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
        - name: path
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: File content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileContent"
    put:
      operationId: saveFileContent
      tags: [files]
      summary: Save file content
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveFileRequest"
      responses:
        "200":
          description: Saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /projects/{projectId}/files/delete:
    post:
      operationId: deleteFile
      tags: [files]
      summary: Delete a file or directory
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeleteFileRequest"
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /projects/{projectId}/files/rename:
    post:
      operationId: renameFile
      tags: [files]
      summary: Rename a file or directory
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RenameFileRequest"
      responses:
        "200":
          description: Renamed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /projects/{projectId}/terminal:
    post:
      operationId: executeCommand
      tags: [terminal]
      summary: Execute a terminal command
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExecuteCommandRequest"
      responses:
        "200":
          description: Command output
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommandOutput"

  /projects/{projectId}/search:
    post:
      operationId: searchFiles
      tags: [files]
      summary: Search files in a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SearchRequest"
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SearchResult"

  /ai/chat:
    post:
      operationId: aiChat
      tags: [ai]
      summary: Chat with AI assistant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiChatRequest"
      responses:
        "200":
          description: AI response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiChatResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message

    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message

    Project:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        type:
          type: string
          enum: [vanilla, react, vue, svelte, node, fullstack]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - type
        - createdAt
        - updatedAt

    CreateProjectRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        type:
          type: string
          enum: [vanilla, react, vue, svelte, node, fullstack]
      required:
        - name
        - type

    FileNode:
      type: object
      properties:
        name:
          type: string
        path:
          type: string
        type:
          type: string
          enum: [file, directory]
        children:
          type: array
          items:
            $ref: "#/components/schemas/FileNode"
        size:
          type: integer
        extension:
          type: string
      required:
        - name
        - path
        - type

    FileContent:
      type: object
      properties:
        path:
          type: string
        content:
          type: string
        language:
          type: string
      required:
        - path
        - content
        - language

    CreateFileRequest:
      type: object
      properties:
        path:
          type: string
        type:
          type: string
          enum: [file, directory]
        content:
          type: string
      required:
        - path
        - type

    SaveFileRequest:
      type: object
      properties:
        path:
          type: string
        content:
          type: string
      required:
        - path
        - content

    DeleteFileRequest:
      type: object
      properties:
        path:
          type: string
      required:
        - path

    RenameFileRequest:
      type: object
      properties:
        oldPath:
          type: string
        newPath:
          type: string
      required:
        - oldPath
        - newPath

    ExecuteCommandRequest:
      type: object
      properties:
        command:
          type: string
        cwd:
          type: string
      required:
        - command

    CommandOutput:
      type: object
      properties:
        stdout:
          type: string
        stderr:
          type: string
        exitCode:
          type: integer
        executionTime:
          type: integer
      required:
        - stdout
        - stderr
        - exitCode

    SearchRequest:
      type: object
      properties:
        query:
          type: string
        caseSensitive:
          type: boolean
          default: false
        useRegex:
          type: boolean
          default: false
        filePattern:
          type: string
      required:
        - query

    SearchResult:
      type: object
      properties:
        file:
          type: string
        line:
          type: integer
        column:
          type: integer
        content:
          type: string
        match:
          type: string
      required:
        - file
        - line
        - content
        - match

    AiChatRequest:
      type: object
      properties:
        message:
          type: string
        context:
          type: string
        projectId:
          type: string
        fileContent:
          type: string
        filePath:
          type: string
        history:
          type: array
          items:
            $ref: "#/components/schemas/ChatMessage"
      required:
        - message

    ChatMessage:
      type: object
      properties:
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
      required:
        - role
        - content

    AiChatResponse:
      type: object
      properties:
        message:
          type: string
        suggestions:
          type: array
          items:
            type: string
      required:
        - message