# lib/api-zod/src/generated/types/aiChatRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Softcurse LiveScriptor IDE API
 * OpenAPI spec version: 0.1.0
 */
import type { ChatMessage } from "./chatMessage";

export interface AiChatRequest {
  message: string;
  context?: string;
  projectId?: string;
  fileContent?: string;
  filePath?: string;
  history?: ChatMessage[];
}