# lib/api-zod/src/generated/types/chatMessage.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Softcurse LiveScriptor IDE API
 * OpenAPI spec version: 0.1.0
 */
import type { ChatMessageRole } from "./chatMessageRole";

export interface ChatMessage {
  role: ChatMessageRole;
  content: string;
}