# lib/api-zod/src/generated/types/commandOutput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Softcurse LiveScriptor IDE API
 * OpenAPI spec version: 0.1.0
 */

export interface CommandOutput {
  stdout: string;
  stderr: string;
  exitCode: number;
  executionTime?: number;
}