# lib/api-zod/src/generated/types/createFileRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Softcurse LiveScriptor IDE API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateFileRequestType } from "./createFileRequestType";

export interface CreateFileRequest {
  path: string;
  type: CreateFileRequestType;
  content?: string;
}