# lib/api-zod/src/generated/types/saveFileRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Softcurse LiveScriptor IDE API
 * OpenAPI spec version: 0.1.0
 */

export interface SaveFileRequest {
  path: string;
  content: string;
}