# lib/api-zod/src/generated/types/searchResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Softcurse LiveScriptor IDE API
 * OpenAPI spec version: 0.1.0
 */

export interface SearchResult {
  file: string;
  line: number;
  column?: number;
  content: string;
  match: string;
}